Cleanup and organize components
This is a first try at some sane organization for a growing code base.
Hopefully, this will also help us test and see what tests are missing.
