+++ borderpatrol-core/src/main/scala/com/lookout/borderpatrol/LoginFilter.scala
@@ -1,6 +1,6 @@
 package com.lookout.borderpatrol
 
-import com.lookout.borderpatrol.session.TokenJson.TokensJson
+import com.lookout.borderpatrol.session.tokens._
 import com.twitter.finagle.http.{Response => FinagleResponse}
 import com.twitter.finagle.{Filter, Service}
 import org.jboss.netty.handler.codec.http.HttpMethod
@@ -20,6 +20,6 @@ class LoginFilter extends Filter[RoutedRequest, AuthResponse, RoutedRequest, Fin
 
   def addTokensToRequest(response: FinagleResponse, request: RoutedRequest): RoutedRequest = {
     val json = response.getContentString()
-    TokensJson(json).foldRight(request)((tokens, req) => req ++ tokens)
+    json.asTokens.foldRight(request)((tokens, req) => req ++ tokens)
   }
 }
