+++ borderpatrol-core/src/main/scala/com/lookout/borderpatrol/session/tokens/TokenState.scala
@@ -1,0 +1,47 @@
+package com.lookout.borderpatrol.session.tokens
+
+import com.lookout.borderpatrol.session.{Token, Tokens, MasterToken, ServiceToken, ServiceTokensBase, ServiceTokens}
+
+import scalaz.State
+
+object TokenState {
+  /**
+   * Add master tokens to the current tokens
+   * @param token A master token
+   * @return
+   */
+  def apply(token: Token): State[Tokens, Tokens] = for {
+    _ <- State.modify((cur: Tokens) => (token, cur) match {
+      case (MasterToken(v), Tokens(MasterToken(w), _)) if v == w => cur
+      case (t: MasterToken, _) => cur.copy(t, cur.services)
+      case (t: ServiceToken, Tokens(_, st)) => cur.copy(cur.master, st + t)
+      case _ => cur
+    })
+    s <- State.get
+  } yield Tokens(s.master, s.services)
+
+  /**
+   * Add service tokens to the current tokens
+   * @param tokens A master token
+   * @return
+   */
+  def apply(tokens: ServiceTokensBase): State[Tokens, Tokens] = for {
+    _ <- State.modify((cur: Tokens) => (tokens, cur) match {
+      case (ServiceTokens(map), _) => cur.copy(cur.master, cur.services ++ tokens)
+      case _ => cur
+    })
+    s <- State.get
+  } yield Tokens(s.master, s.services)
+
+  /**
+   * Converge new tokens with the current tokens
+   * @param tokens A master token
+   * @return
+   */
+  def apply(tokens: Tokens): State[Tokens, Tokens] = for {
+    _ <- State.modify((cur: Tokens) => (tokens, cur) match {
+      case (Tokens(mt, st), _) => cur.copy(mt, cur.services ++ st)
+    })
+    s <- State.get
+  } yield Tokens(s.master, s.services)
+}
